Update index.html
Nothing, just reorganized to find the icon for the "hambuger icon" that toggles the sidebar. I was trying to change that icon because I mistaked it as a  collapsed nav bar and when the screeen is minimized, the icon changed from left to right location. Not staying fixed in the web browser. Again, it's nothing becasue there's no changes.